Fix failing tests.
