+++ sources/mandatory/signals/hooks_shell.c
@@ -6,18 +6,18 @@
 /*   By: lpaulo-m <lpaulo-m@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/10 15:51:17 by lpaulo-m          #+#    #+#             */
-/*   Updated: 2022/10/11 15:02:15 by lpaulo-m         ###   ########.fr       */
+/*   Updated: 2022/10/16 14:43:55 by lpaulo-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <minishell.h>
 
-static void	set_interrupt_signal_hook(void)
+void	set_interrupt_signal_hook(void)
 {
 	set_signal_hook(signal_action(), handle_interrupt_signal, SIGINT);
 }
 
-static void	set_quit_signal_hook(void)
+void	set_quit_signal_hook(void)
 {
 	set_signal_hook(signal_action(), SIG_IGN, SIGQUIT);
 }
@@ -26,9 +26,4 @@ void	set_interactive_shell_hooks(void)
 {
 	set_interrupt_signal_hook();
 	set_quit_signal_hook();
-}
-
-void	disable_interrupt_signal(void)
-{
-	set_signal_hook(signal_action(), SIG_IGN, SIGINT);
 }
